Add Imunify360 check to Security Advisor.
Case LC-10352: This commit is a squash of the Imunify360 check
changes from LC story 10290.

The check covers the following items:

- Does this server have an Imunify360 license?

- Is Imunify360 actually installed on this server?

If the answer to either of the above is "no," then a suggestion
is displayed. We are only providing a link to the cPanel store
currently and have not integrated the purchase process into WHM. A
future story is expected to cover the tighter integration of the
purchase process, similar to what's done with KernelCare.

Requirements:

- The new check will only work on v80+, and we only intend to update
the submodule commit for v80 and greater. If the submodule commit
does get bumped up on an older version, the Security Advisor will
still function, but the new check will not be included.